+++ lib/util/mixins/use_sprite_animation.dart
@@ -1,6 +1,8 @@
+import 'dart:async';
 import 'dart:ui';
 
 import 'package:bonfire/base/game_component.dart';
+import 'package:bonfire/objects/animated_object_once.dart';
 import 'package:bonfire/util/extensions/extensions.dart';
 import 'package:flame/components.dart';
 
@@ -19,16 +21,29 @@ mixin UseSpriteAnimation on GameComponent {
   /// Animation that will be drawn on the screen.
   SpriteAnimation? animation;
 
+  /// Offset of the render animation.
+  Vector2 animationOffset = Vector2.zero();
+
+  /// Size animation. if null use component size
+  Vector2? animationSize;
+  AnimatedObjectOnce? _fastAnimation;
+  Vector2 _fastAnimOffset = Vector2.zero();
+
   @override
   void render(Canvas canvas) {
     super.render(canvas);
+
     if (isVisible) {
-      animation?.getSprite().renderWithOpacity(
-            canvas,
-            this.position,
-            this.size,
-            opacity: opacity,
-          );
+      if (_fastAnimation != null) {
+        _fastAnimation?.render(canvas);
+      } else {
+        animation?.getSprite().renderWithOpacity(
+              canvas,
+              position + animationOffset,
+              animationSize ?? size,
+              opacity: opacity,
+            );
+      }
     }
   }
 
@@ -36,7 +51,35 @@ mixin UseSpriteAnimation on GameComponent {
   void update(double dt) {
     super.update(dt);
     if (this.isVisible) {
+      _fastAnimation?.position = position + _fastAnimOffset;
+      _fastAnimation?.opacity = opacity;
+      _fastAnimation?.isFlipHorizontal = isFlipHorizontal;
+      _fastAnimation?.isFlipVertical = isFlipVertical;
+      _fastAnimation?.update(dt);
       animation?.update(dt);
     }
+  }
+
+  /// Method used to play animation once time
+  Future playSpriteAnimationOnce(
+    FutureOr<SpriteAnimation> animation, {
+    Vector2? size,
+    Vector2? offset,
+    VoidCallback? onFinish,
+    VoidCallback? onStart,
+  }) async {
+    _fastAnimOffset = offset ?? Vector2.zero();
+    final anim = AnimatedObjectOnce(
+      position: position + _fastAnimOffset,
+      size: size ?? this.size,
+      animation: animation,
+      onStart: onStart,
+      onFinish: () {
+        onFinish?.call();
+        _fastAnimation = null;
+      },
+    )..gameRef = gameRef;
+    await anim.onLoad();
+    _fastAnimation = anim;
   }
 }